+++ Back-end/src/main/java/it/ErdisonDosti/service/ServiceDTO.java
@@ -12,11 +12,11 @@ public interface ServiceDTO<DTO> {
 	
 	public Iterable<DTO> getAll();
 
-	public DTO read(long id);
+	public DTO read(Integer id);
 	
 	public DTO insert (DTO dto);
 	
 	public DTO update (DTO dto);
 	
-	public void delete (long id);
+	public void delete (Integer id);
 }
